+++ aiocache/decorators.py
@@ -38,8 +38,11 @@ class cached:
     :param key: str value to set as key for the function return. Takes precedence over
         key_builder param. If key and key_builder are not passed, it will use module_name
         + function_name + args + kwargs
+    :param namespace: string to use as default prefix for the key used in all operations of
+        the backend. Default is None
     :param key_builder: Callable that allows to build the function dynamically. It receives
         the function plus same args and kwargs passed to the function.
+        This behavior is necessarily different than ``BaseCache.build_key()``
     :param cache: cache class to use when calling the ``set``/``get`` operations.
         Default is :class:`aiocache.SimpleMemoryCache`.
     :param serializer: serializer instance to use when calling the ``dumps``/``loads``.
@@ -58,6 +61,7 @@ class cached:
         self,
         ttl=SENTINEL,
         key=None,
+        namespace=None,
         key_builder=None,
         cache=Cache.MEMORY,
         serializer=None,
@@ -75,16 +79,21 @@ class cached:
 
         self._cache = cache
         self._serializer = serializer
+        self._namespace = namespace
         self._plugins = plugins
         self._kwargs = kwargs
 
     def __call__(self, f):
         if self.alias:
             self.cache = caches.get(self.alias)
+            for arg in ("serializer", "namespace", "plugins"):
+                if getattr(self, f'_{arg}', None) is not None:
+                    logger.warning(f"Using cache alias; ignoring '{arg}' argument.")
         else:
             self.cache = _get_cache(
                 cache=self._cache,
                 serializer=self._serializer,
+                namespace=self._namespace,
                 plugins=self._plugins,
                 **self._kwargs,
             )
@@ -168,6 +177,11 @@ class cached_stampede(cached):
         key_from_attr param. If key and key_from_attr are not passed, it will use module_name
         + function_name + args + kwargs
     :param key_from_attr: str arg or kwarg name from the function to use as a key.
+    :param namespace: string to use as default prefix for the key used in all operations of
+        the backend. Default is None
+    :param key_builder: Callable that allows to build the function dynamically. It receives
+        the function plus same args and kwargs passed to the function.
+        This behavior is necessarily different than ``BaseCache.build_key()``
     :param cache: cache class to use when calling the ``set``/``get`` operations.
         Default is :class:`aiocache.SimpleMemoryCache`.
     :param serializer: serializer instance to use when calling the ``dumps``/``loads``.
@@ -250,8 +264,11 @@ class multi_cached:
 
     :param keys_from_attr: arg or kwarg name from the function containing an iterable to use
         as keys to index in the cache.
-    :param key_builder: Callable that allows to change the format of the keys before storing.
-        Receives the key the function and same args and kwargs as the called function.
+    :param namespace: string to use as default prefix for the key used in all operations of
+        the backend. Default is None
+    :param key_builder: Callable that allows to build the function dynamically. It receives
+        the function plus same args and kwargs passed to the function.
+        This behavior is necessarily different than ``BaseCache.build_key()``
     :param ttl: int seconds to store the keys. Default is 0 which means no expiration.
     :param cache: cache class to use when calling the ``multi_set``/``multi_get`` operations.
         Default is :class:`aiocache.SimpleMemoryCache`.
@@ -268,6 +285,7 @@ class multi_cached:
     def __init__(
         self,
         keys_from_attr,
+        namespace=None,
         key_builder=None,
         ttl=SENTINEL,
         cache=Cache.MEMORY,
@@ -284,16 +302,21 @@ class multi_cached:
 
         self._cache = cache
         self._serializer = serializer
+        self._namespace = namespace
         self._plugins = plugins
         self._kwargs = kwargs
 
     def __call__(self, f):
         if self.alias:
             self.cache = caches.get(self.alias)
+            for arg in ("serializer", "namespace", "plugins"):
+                if getattr(self, f'_{arg}', None) is not None:
+                    logger.warning(f"Using cache alias; ignoring '{arg}' argument.")
         else:
             self.cache = _get_cache(
                 cache=self._cache,
                 serializer=self._serializer,
+                namespace=self._namespace,
                 plugins=self._plugins,
                 **self._kwargs,
             )