+++ tests/acceptance/test_decorators.py
@@ -5,6 +5,7 @@ from unittest import mock
 import pytest
 
 from aiocache import cached, cached_stampede, multi_cached
+from aiocache.base import _ensure_key
 from ..utils import Keys
 
 
@@ -47,6 +48,30 @@ class TestCached:
 
         await fn("self", 1, 3)
         assert await cache.exists(build_key(fn, "self", 1, 3)) is True
+
+    async def test_cached_without_namespace(self, cache):
+        """Default cache key is created when no namespace is provided"""
+        @cached(namespace=None)
+        async def fn():
+            return "1"
+
+        await fn()
+        decorator = cached(namespace=None)
+        key = decorator.get_cache_key(fn, args=(), kwargs={})
+        assert await cache.exists(key, namespace=None) is True
+
+    async def test_cached_with_namespace(self, cache):
+        """Cache key is prefixed with provided namespace"""
+        key_prefix = "test"
+
+        @cached(namespace=key_prefix)
+        async def ns_fn():
+            return "1"
+
+        await ns_fn()
+        decorator = cached(namespace=key_prefix)
+        key = decorator.get_cache_key(ns_fn, args=(), kwargs={})
+        assert await cache.exists(key, namespace=key_prefix) is True
 
 
 class TestCachedStampede:
@@ -113,17 +138,16 @@ class TestMultiCachedDecorator:
         assert await cache.exists(Keys.KEY) is True
 
     async def test_multi_cached_key_builder(self, cache):
-        # TODO(PY311): Remove str() calls
         def build_key(key, f, self, keys, market="ES"):
-            return "{}_{}_{}".format(f.__name__, str(key), market)
+            return "{}_{}_{}".format(f.__name__, _ensure_key(key), market)
 
         @multi_cached(keys_from_attr="keys", key_builder=build_key)
         async def fn(self, keys, market="ES"):
             return {Keys.KEY: 1, Keys.KEY_1: 2}
 
         await fn("self", keys=[Keys.KEY, Keys.KEY_1])
-        assert await cache.exists("fn_" + str(Keys.KEY) + "_ES") is True
-        assert await cache.exists("fn_" + str(Keys.KEY_1) + "_ES") is True
+        assert await cache.exists("fn_" + _ensure_key(Keys.KEY) + "_ES") is True
+        assert await cache.exists("fn_" + _ensure_key(Keys.KEY_1) + "_ES") is True
 
     async def test_fn_with_args(self, cache):
         @multi_cached("keys")