+++ tests/ut/test_base.py
@@ -4,7 +4,7 @@ from unittest.mock import ANY, AsyncMock, MagicMock, patch
 
 import pytest
 
-from aiocache.base import API, BaseCache, _Conn
+from aiocache.base import API, BaseCache, _Conn, _ensure_key
 from ..utils import Keys
 
 
@@ -205,8 +205,7 @@ class TestBaseCache:
 
     @pytest.mark.parametrize(
         "namespace, expected",
-        # TODO(PY311): Remove str()
-        ([None, "test" + str(Keys.KEY)], ["", str(Keys.KEY)], ["my_ns", "my_ns" + str(Keys.KEY)]),  # type: ignore[attr-defined]  # noqa: B950
+        ([None, "test" + _ensure_key(Keys.KEY)], ["", _ensure_key(Keys.KEY)], ["my_ns", "my_ns" + _ensure_key(Keys.KEY)]),  # type: ignore[attr-defined]  # noqa: B950
     )
     def test_build_key(self, set_test_namespace, base_cache, namespace, expected):
         assert base_cache.build_key(Keys.KEY, namespace=namespace) == expected
@@ -214,6 +213,106 @@ class TestBaseCache:
     def test_alt_build_key(self):
         cache = BaseCache(key_builder=lambda key, namespace: "x")
         assert cache.build_key(Keys.KEY, "namespace") == "x"
+
+    @pytest.fixture
+    def alt_base_cache(self, init_namespace="test"):
+        """Custom key_builder for cache"""
+        def build_key(key, namespace=None):
+            ns = namespace if namespace is not None else ""
+            sep = ":" if namespace else ""
+            return f"{ns}{sep}{_ensure_key(key)}"
+
+        cache = BaseCache(key_builder=build_key, namespace=init_namespace)
+        return cache
+
+    @pytest.mark.parametrize(
+        "namespace, expected",
+        ([None, _ensure_key(Keys.KEY)], ["", _ensure_key(Keys.KEY)], ["my_ns", "my_ns:" + _ensure_key(Keys.KEY)]),  # type: ignore[attr-defined]  # noqa: B950
+    )
+    def test_alt_build_key_override_namespace(self, alt_base_cache, namespace, expected):
+        """Custom key_builder overrides namespace of cache"""
+        cache = alt_base_cache
+        assert cache.build_key(Keys.KEY, namespace=namespace) == expected
+
+    @pytest.mark.parametrize(
+        "init_namespace, expected",
+        ([None, _ensure_key(Keys.KEY)], ["", _ensure_key(Keys.KEY)], ["test", "test:" + _ensure_key(Keys.KEY)]),  # type: ignore[attr-defined]  # noqa: B950
+    )
+    async def test_alt_build_key_default_namespace(
+            self, init_namespace, alt_base_cache, expected):
+        """Custom key_builder for cache with or without namespace specified.
+
+        Cache member functions that accept a ``namespace`` parameter
+        should default to using ``self.namespace`` if the ``namespace``
+        argument is ``None``.
+
+        This enables a cache to correctly build keys when the cache is
+        initialized with both a ``namespace`` and a ``key_builder``,
+        even when that cache is supplied to a lock or to a decorator
+        using the ``alias`` argument.
+        """
+        cache = alt_base_cache
+        cache.namespace = init_namespace
+
+        # Verify that private members are called with the correct ns_key
+        await self._assert_add__alt_build_key_default_namespace(cache, expected)
+        await self._assert_get__alt_build_key_default_namespace(cache, expected)
+        await self._assert_multi_get__alt_build_key_default_namespace(cache, expected)
+        await self._assert_set__alt_build_key_default_namespace(cache, expected)
+        await self._assert_multi_set__alt_build_key_default_namespace(cache, expected)
+        await self._assert_exists__alt_build_key_default_namespace(cache, expected)
+        await self._assert_increment__alt_build_key_default_namespace(cache, expected)
+        await self._assert_delete__alt_build_key_default_namespace(cache, expected)
+        await self._assert_expire__alt_build_key_default_namespace(cache, expected)
+
+    async def _assert_add__alt_build_key_default_namespace(self, cache, expected):
+        with patch.object(cache, "_add", autospec=True) as _add:
+            await cache.add(Keys.KEY, "value")
+            _add.assert_called_once_with(expected, "value", _conn=None, ttl=None)
+
+    async def _assert_get__alt_build_key_default_namespace(self, cache, expected):
+        with patch.object(cache, "_get", autospec=True) as _get:
+            await cache.get(Keys.KEY)
+            _get.assert_called_once_with(
+                expected, _conn=None, encoding=cache.serializer.encoding)
+
+    async def _assert_multi_get__alt_build_key_default_namespace(self, cache, expected):
+        with patch.object(cache, "_multi_get", autospec=True) as _multi_get:
+            await cache.multi_get([Keys.KEY])
+            _multi_get.assert_called_once_with(
+                [expected], _conn=None, encoding=cache.serializer.encoding)
+
+    async def _assert_set__alt_build_key_default_namespace(self, cache, expected):
+        with patch.object(cache, "_set", autospec=True) as _set:
+            await cache.set(Keys.KEY, "value")
+            _set.assert_called_once_with(
+                expected, "value", _conn=None, ttl=None, _cas_token=None)
+
+    async def _assert_multi_set__alt_build_key_default_namespace(self, cache, expected):
+        with patch.object(cache, "_multi_set", autospec=True) as _multi_set:
+            await cache.multi_set([(Keys.KEY, "value")])
+            _multi_set.assert_called_once_with(
+                [(expected, "value")], _conn=None, ttl=None)
+
+    async def _assert_exists__alt_build_key_default_namespace(self, cache, expected):
+        with patch.object(cache, "_exists", autospec=True) as _exists:
+            await cache.exists(Keys.KEY)
+            _exists.assert_called_once_with(expected, _conn=None)
+
+    async def _assert_increment__alt_build_key_default_namespace(self, cache, expected):
+        with patch.object(cache, "_increment", autospec=True) as _increment:
+            await cache.increment(Keys.KEY)
+            _increment.assert_called_once_with(expected, delta=1, _conn=None)
+
+    async def _assert_delete__alt_build_key_default_namespace(self, cache, expected):
+        with patch.object(cache, "_delete", autospec=True) as _delete:
+            await cache.delete(Keys.KEY)
+            _delete.assert_called_once_with(expected, _conn=None)
+
+    async def _assert_expire__alt_build_key_default_namespace(self, cache, expected):
+        with patch.object(cache, "_expire", autospec=True) as _expire:
+            await cache.expire(Keys.KEY, 0)
+            _expire.assert_called_once_with(expected, 0, _conn=None)
 
     async def test_add_ttl_cache_default(self, base_cache):
         with patch.object(base_cache, "_add", autospec=True) as m: